requirements.txt
requests
BeautifulSoup4
langchain_text_splitters
langchain
sentence_transformers
faiss
langchain_community
lanchain-chroma
groq
python-dotenv
langchain-core
langchain-groq
